7/02/2006 - giwo (0.98-3.6h) [3.7 RC] 	Removed door.cpp from the project. 	Modified the JS UseDoor() function to call the JS Event onUse() based upon the type of the item passed in. 	Fixed an issue causing AreaCharacterFunction and AreaItemFunction to return the number of objects they encountered 		rather than the number of times the function returns a "true" value. 	Modified TriggerEvent() to return true or false based upon the return value of the function. 	Moved doors out to JS.

+++ source/cHTMLSystem.cpp
@@ -218,7 +218,7 @@ void cHTMLTemplate::Process( void )
 	// Timestamp
 	time_t currTime;
 	time( &currTime );
-	UString timestamp = UString::number( currTime );
+	UString timestamp = UString::number( static_cast<size_t>(currTime) );
 	Pos = ParsedContent.find( "%tstamp" ); 
 	while( Pos != std::string::npos )
 	{
